bug 1334428, cope with nightly buildids that don't have all platforms. (#52)
Should also help with bug 1331707 and give the tasks more meaningful symbols in treeherder.

Since not every build has all platforms, we need to generate a list of relevant builds and
work on that, which means rearranging the internal loop.  Since submitting in the previous
place would then break the update_number, which makes it into the final MAR, move the
task submission out to the end, so we can generate the tasks in the proper order.

It means each Update-\d group in treeherder is for a particular 'from' generation, making
it hopefully easier to track what's going on.